cleaned up navbar file by seperating out elements
